--- a/libnestutil/propagator_stability.cpp
+++ b/libnestutil/propagator_stability.cpp
@@ -28,29 +28,58 @@
 // Includes from libnestutil:
 #include "numerics.h"
 
-double
-propagator_31( double tau_syn, double tau, double C, double h )
+propagator::propagator( )
+  : alpha_( 0 )
+  , beta_ ( 0 )
+  , gamma_ ( 0 )
+  , gamma_sq_ ( 0 )
 {
-  const double P31_linear = 1 / ( 3. * C * tau * tau ) * h * h * h * ( tau_syn - tau ) * std::exp( -h / tau );
-  const double P31 = 1 / C
-    * ( std::exp( -h / tau_syn ) * numerics::expm1( -h / tau + h / tau_syn ) / ( tau / tau_syn - 1 ) * tau
-      - h * std::exp( -h / tau_syn ) )
-    / ( -1 - -tau / tau_syn ) * tau;
-  const double P31_singular = h * h / 2 / C * std::exp( -h / tau );
-  const double dev_P31 = std::abs( P31 - P31_singular );
+}
 
-  if ( tau == tau_syn or ( std::abs( tau - tau_syn ) < 0.1 and dev_P31 > 2 * std::abs( P31_linear ) ) )
-  {
-    return P31_singular;
-  }
-  else
-  {
-    return P31;
-  }
+void propagator::calculate_constants( double tau_syn, double tau, double c )
+{
+  alpha_ = 1 / ( 3. * c * tau * tau ) * ( tau_syn - tau );
+  beta_ = tau_syn  * tau / (tau - tau_syn);
+  gamma_ = 1 / c * beta_;
+  gamma_sq_ = 1 / c / ( ( 1 / tau_syn - 1 / tau ) * ( 1 / tau_syn - 1 / tau ) );
 }
 
 double
-propagator_32( double tau_syn, double tau, double C, double h )
+propagator::propagator_31( double tau_syn, double tau, double C, double h ) const
+{
+  const double exp_h_tau_syn = std::exp( -h / tau_syn );
+  const double expm1_h_tau = numerics::expm1( -h / tau + h / tau_syn );
+
+
+  //const double P31_linear = 1 / ( 3. * C * tau * tau ) * h * h * h * ( tau_syn - tau ) * std::exp( -h / tau );
+
+  //const double P31 = 1 / C
+  //  * ( std::exp( -h / tau_syn ) * numerics::expm1( -h / tau + h / tau_syn ) / ( tau / tau_syn - 1 ) * tau
+  //    - h * std::exp( -h / tau_syn ) )
+  //  / ( -1 - -tau / tau_syn ) * tau;
+
+  const double P31 = gamma_ * ( beta_ * exp_h_tau_syn * expm1_h_tau - h * exp_h_tau_syn);
+
+  //const double P31_singular = h * h / 2 / C * std::exp( -h / tau );
+  //const double dev_P31 = std::abs( P31 - P31_singular );
+
+  if ( tau == tau_syn or std::abs( tau - tau_syn ) < 0.1 )
+  {
+    const double P31_linear = alpha_ * h * h * h * std::exp( -h / tau );
+    const double P31_singular = h * h / 2 / C * std::exp( -h / tau );
+    const double dev_P31 = std::abs( P31 - P31_singular );
+
+    if ( dev_P31 > 2 * std::abs( P31_linear ) )
+    {
+      return P31_singular;
+    }
+  }
+
+  return P31;
+}
+
+double
+propagator::propagator_32( double tau_syn, double tau, double C, double h ) const
 {
   const double P32_linear = 1 / ( 2. * C * tau * tau ) * h * h * ( tau_syn - tau ) * std::exp( -h / tau );
   const double P32_singular = h / C * std::exp( -h / tau );
@@ -67,4 +96,16 @@ propagator_32( double tau_syn, double tau, double C, double h )
   {
     return P32;
   }
+}
+
+double
+propagator_31( double tau_syn, double tau, double C, double h )
+{
+  return 1.;
+}
+
+double
+propagator_32( double tau_syn, double tau, double C, double h )
+{
+  return 1.;
 }