--- a/libnestutil/propagator_stability.cpp
+++ b/libnestutil/propagator_stability.cpp
@@ -29,29 +29,28 @@
 #include "numerics.h"
 
 propagator::propagator()
-  : alpha_( 0 )
-  , beta_ ( 0 )
-  , gamma_ ( 0 )
-  , tau_syn_ ( 0 )
-  , tau_ ( 0 )
-  , c_ ( 0 )
+  : alpha_( 0.0 )
+  , beta_( 0.0 )
+  , gamma_( 0.0 )
+  , tau_syn_( 0.0 )
+  , tau_( 0.0 )
+  , c_( 0.0 )
 {
 }
 
 propagator::propagator( double tau_syn, double tau, double c )
-  : tau_syn_ ( tau_syn )
-  , tau_ ( tau )
-  , c_ ( c )
+  : tau_syn_( tau_syn )
+  , tau_( tau )
+  , c_( c )
 {
-  alpha_ = 1 / ( c * tau * tau ) * ( tau_syn - tau );
-  beta_ = tau_syn  * tau / ( tau - tau_syn );
-  gamma_ = beta_ / c;
+  update_constants( tau_syn, tau, c );
 }
 
-void propagator::update_constants( double tau_syn, double tau, double c )
+void
+propagator::update_constants( double tau_syn, double tau, double c )
 {
   alpha_ = 1 / ( c * tau * tau ) * ( tau_syn - tau );
-  beta_ = tau_syn  * tau / ( tau - tau_syn );
+  beta_ = tau_syn * tau / ( tau - tau_syn );
   gamma_ = beta_ / c;
 
   tau_syn_ = tau_syn;