--- a/libnestutil/propagator_stability.cpp
+++ b/libnestutil/propagator_stability.cpp
@@ -100,7 +100,7 @@ PropagatorAlpha::evaluate( double h ) const
   double P31 = gamma_ * exp_h_tau_syn * ( beta_ * expm1_h_tau - h );
   double P32 = gamma_ * exp_h_tau_syn * expm1_h_tau;
 
-  if ( tau_ == tau_syn_ or std::abs( tau_ - tau_syn_ ) < 0.1 )
+  if ( std::abs( tau_ - tau_syn_ ) < 0.1 )
   {
     const double exp_h_tau = std::exp( -h / tau_ );