This PR follows up the discussion in #2294 and #2017. I have created two propagator classes, `PropagateExp` and `PropagateAlpha` for calculating propagators for models with exponential and alpha postsynaptic currents. Both classes have an `evaluate` function where the exp version just calculates `P31`, and the alpha version calculates `P31` and `P32` simultaneously. This way, the exponentials are calculated fewer times than what we do currently. I have not been able to get the simulation time for `iaf_psc_alpha_ps` to precisely match `iaf_psc_alpha_canon`, but the times are closer than with master, see results below.

With this new version, getting P31 and P32 is handled slightly differently than before:
Old version
```
V_.P31_ = propagator_31( P_.tau_syn_, tau_m, P_.c_m_, h );
V_.P32_ = propagator_32( P_.tau_syn_, tau_m, P_.c_m_, h );
```
New version exp model
```
V_.P21_ = PropagatorExp( P_.tau_syn_, tau_m, P_.c_m_ ).evaluate( h );
```
New version alpha model
```
std::tie( V_.P31_, V_.P32_ ) = PropagatorAlpha( P_.tau_syn_, tau_m_, P_.c_m_ ).evaluate( h );
```

Here are some timing results. All measured on my computer.
| model  | Simulation time for `brunel_ps.sli` | Time `ArtificialSyncrony.sli`
| ------------- | ------------- | ------------- |
| `iaf_psc_alpha_ps` with `propagator` class  | 132.82 sec  | 7 min 23 sec |
| master version of `iaf_psc_alpha_ps`  | 222.86 sec  | 9 min 18 sec |
| master version of `iaf_psc_alpha_canon`  | 95.03 sec | 6 min 33 sec |

`iaf_psc_alpha_canon` does not use `propagate_31` and `propagate_32` in it's [`propogate_` function](https://github.com/nest/nest-simulator/blob/master/models/iaf_psc_alpha_canon.cpp#L473). Using canon's version in the ps model decrease the time further, see comment [here](https://github.com/nest/nest-simulator/pull/2294#issuecomment-1065174892).

@heplesser